test/resnet/requirements.txt
absl-py==2.0.0
astunparse==1.6.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
cachetools==5.3.2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.2.0
click==8.0.3
cloud-init==23.2.1
cloud-tpu-client==0.10
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
Cython==0.29.28
dbus-python==1.2.18
dill==0.3.7
distlib==0.3.7
distro==1.7.0
distro-info===1.1build1
dm-tree==0.1.8
filelock==3.13.0
flatbuffers==23.5.26
fsspec==2023.10.0
gast==0.5.4
google-api-core==1.34.0
google-api-python-client==1.8.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
gviz-api==1.10.0
h5py==3.10.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keras-nightly==3.0.0.dev2023102903
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.6
libtpu-nightly==0.1.dev20230825+default
lmdb==1.4.1
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ml-dtypes==0.3.1
more-itertools==8.10.0
mpmath==1.3.0
namex==0.0.7
netifaces==0.11.0
networkx==3.2.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.10.0
protobuf==3.20.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
Pygments==2.16.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu1
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rsa==4.9
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sos==4.4
ssh-import-id==5.11
sympy==1.12
systemd-python==234
tb-nightly==2.15.0a20231023
tbp-nightly==2.15.0a20231028
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
tf-estimator-nightly==2.14.0.dev2023080308
tf-nightly==2.16.0.dev20231026
torch==2.1.0
torch-xla==2.1.0
torchvision==0.16.0
triton==2.1.0
Twisted==22.1.0
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.5
virtualenv==20.24.2
wadllib==1.3.6
Werkzeug==3.0.1
wrapt==1.14.1
zipp==1.0.0
zope.interface==5.4.0
